+++ CHANGELOG.md
@@ -3,7 +3,14 @@
 Adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [0.13.0] - unreleased
+## [0.13.1] - 2021-09-12
+
+### Fixed
+
+- `CompressionFormat.strip`
+- Deprecation warning in test
+
+## [0.13.0] - 2021-09-11
 
 ### Added